Set C binding for unix_fsync_job in its own C file
