+++ src/unix/jbuild
@@ -209,6 +209,7 @@
     unix_tcgetattr_job
     unix_tcsetattr_job
     windows_is_socket
+    windows_fsync_job
     windows_not_available
     unix_not_available
   ))
