+++ src/unix/lwt_unix_windows.h
@@ -20,49 +20,6 @@
  */
 
 /* Windows version of stubs. */
-
-
-
-/* +-----------------------------------------------------------------+
-   | JOB: fsync                                                      |
-   +-----------------------------------------------------------------+ */
-
-struct job_fsync {
-    struct lwt_unix_job job;
-    HANDLE handle;
-    DWORD error_code;
-};
-
-static void worker_fsync(struct job_fsync *job)
-{
-    if (!FlushFileBuffers(job->handle)) job->error_code = GetLastError();
-}
-
-static value result_fsync(struct job_fsync *job)
-{
-    DWORD error = job->error_code;
-    if (error) {
-        lwt_unix_free_job(&job->job);
-        win32_maperr(error);
-        uerror("fsync", Nothing);
-    }
-    lwt_unix_free_job(&job->job);
-    return Val_unit;
-}
-
-CAMLprim value lwt_unix_fsync_job(value val_fd)
-{
-    struct filedescr *fd = (struct filedescr *)Data_custom_val(val_fd);
-    if (fd->kind != KIND_HANDLE) {
-        caml_invalid_argument("Lwt_unix.fsync");
-    } else {
-        LWT_UNIX_INIT_JOB(job, fsync, 0);
-        job->handle = fd->fd.handle;
-        job->error_code = 0;
-        return lwt_unix_alloc_job(&(job->job));
-    }
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: system                                                     |
    +-----------------------------------------------------------------+ */
@@ -112,8 +69,3 @@ CAMLprim value lwt_unix_system_job(value cmdline)
         return lwt_unix_alloc_job(&(job->job));
     }
 }
-
-/* +-----------------------------------------------------------------+
-   | Unavailable primitives                                          |
-   +-----------------------------------------------------------------+ */
-
